+++ ext_emconf.php
@@ -2,7 +2,7 @@
 
 $EM_CONF[$_EXTKEY] = [
 	'title' => 'PDF ViewHelpers',
-	'description' => 'Provides various Fluid ViewHelpers to create PDF documents. It is possible to use existing PDFs as template and extend them using these ViewHelpers. Under the hood pdfviewhelpers uses TCPDF and FPDI.',
+	'description' => 'Provides various Fluid ViewHelpers to create PDF documents. Under the hood pdfviewhelpers uses TCPDF and FPDI.',
 	'category' => 'fe',
 	'author' => 'Markus Mächler, Esteban Marin',
 	'author_email' => 'markus.maechler@bithost.ch, esteban.marin@bithost.ch',
@@ -12,7 +12,7 @@ $EM_CONF[$_EXTKEY] = [
 	'uploadfolder' => '0',
 	'createDirs' => '',
 	'clearCacheOnLoad' => 0,
-	'version' => '1.6.0',
+	'version' => '2.0.0',
 	'constraints' => [
 		'depends' => [
 			'typo3' => '7.6.0-9.5.99',